requirements.txt
# Core data processing
pandas>=2.1.0
numpy>=1.24.0
dask[complete]>=2023.9.0
pyarrow>=14.0.0

# Data acquisition
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Data validation
great-expectations>=0.17.0
pydantic>=2.5.0

# Database (optional)
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Excel/CSV handling
openpyxl>=3.1.0
xlrd>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Jupyter
jupyterlab>=4.0.0
ipykernel>=6.27.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.66.0
click>=8.1.0

# Logging
loguru>=0.7.0

# Development
black>=23.0.0
isort>=5.13.0
flake8>=6.1.0
mypy>=1.7.0
